refactor: refactor mock data for tests
